pyproject.toml
[project]
name = "registration-verification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.46",
    "flask>=3.1.2",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pymongo>=4.15.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "numpy>=1.26.0",
    "opencv-python>=4.8.0",
]
